requirements.txt
requests>=2.32.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
openai>=1.0.0
azure-search-documents>=11.6.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
pypdf>=4.0.0
pymupdf>=1.24.0
pytest>=8.0.0
httpx>=0.27.0
ddgs>=9.0.0
python-docx>=1.0.0
reportlab>=4.0.0
beautifulsoup4>=4.12.0

# ML / ranking experiments
mlflow>=2.12.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
xgboost>=2.0.0
